pipeline/pipeline_prod/requirements.txt
functions-framework==3.*
pandas
db-dtypes==1.1.1
google-cloud-storage
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.19.1
numpy
scikit-learn==1.3.1
scipy==1.11.2
matplotlib==3.8.0
seaborn==0.12.2
Flask
google-auth